+++ ChangeLog.md
@@ -5,8 +5,9 @@
 (2026-xx-yy)
 
 - added `scheme` and `urlscheme` as synonyms for `url`
+- added `uti` as synonym for `type`
 - added `app utis` and `app urls` as synonyms for `app types` and `app schemes`
-- added `-b` as short flag for `--bundle-id`
+- added `-b` as short flag for `--bundle-id` in `type` and `url`
 
 ## v1.5